Feature/as 3105 create crud endpoints for ivpr (#29)
* AS-3105: Implement InventoryItemDetailDto.

* AS-3105: Implement InventoryItemDetail.

* AS-3105: Implement InventoryItemDetailRsrc.

* AS-3105: Implement InventoryItemDetailList.

* AS-3105: Implement InventoryItemDetailListRsrc.

* AS-3105: Implement InventoryItemDetailFactory.

* AS-3105: Implement InventoryItemDetailEndpoints.

* AS-3105: Fix documentation for BenchmarkPerUnitEndpoints.

* AS-3105: Fix documentation for FairMarketValueEndpoints.

* AS-3105: Fix documentation for InventoryItemDetailRsrcFactory.

* AS-3105: Implement InventoryItemDetailMapper.

* AS-3105: Implement InventoryItemDetailDao.

* AS-3105: Implement InventoryItemDetailMapper.

* AS-3105: Implement InventoryItemDetailDaoImpl.

* AS-3105: Implement InventoryItemDetailService.

* AS-3105: Implement InventoryItemDetailRsrcConstraints.

* AS-3105: Implement InventoryItemDetailServiceImpl.

* AS-3105: Implement InventoryItemDetailEndpointsImpl.

* AS-3105: Implement InventoryItemDetailEndpointsImpl.

* AS-3105: Fix SQL error.

* AS-3105: Fix the equals logic for BigDecimal.

* AS-3105: Change in_version_id datatype from numeric to bigint.

* AS-3105: Fix column name for local_statement_a_receivd_date.

* AS-3105: Fix call to farms_staging_pkg.insert_z99.

* AS-3105: Fix call to farms_staging_pkg.insert_z99.

* AS-3105: Fix call to farms_staging_pkg.insert_z01.

* AS-3105: Fix call to farms_staging_pkg.insert_z02.

* AS-3105: Fix setInt and setShort.

* AS-3105: Fix call to farms_staging_pkg.insert_z03.

* AS-3105: Fix call to farms_staging_pkg.insert_z04.

* AS-3105: Fix call to farms_staging_pkg.insert_z28.

* AS-3105: Fix call to farms_staging_pkg.insert_z21.

* AS-3105: Fix in_participant_pin datatype.

* AS-3105: Fix person_id datatype.

* AS-3105: Fix v_program_year datatype.

* AS-3105: Fix pyv_id datatype.

* AS-3105: Fix fo_id datatype.

* AS-3105: Fix py_id datatype.

* AS-3105: Fix scenario_id datatype.

+++ farms-liquibase/scripts/01_00_xx/01_00_02/06/ddl/routines/farms_import_pkg.participant.sql
@@ -1,6 +1,6 @@
 create or replace function farms_import_pkg.participant(
-    in in_version_id numeric,
-    in in_participant_pin numeric,
+    in in_version_id bigint,
+    in in_participant_pin integer,
     out out_agristability_client_id farms.farm_agristability_clients.agristability_client_id%type,
     in in_user varchar,
     inout in_out_activity numeric,
@@ -118,8 +118,8 @@ declare
     p1_msg varchar(4000) := null;
     p2_msg varchar(4000) := null;
 
-    person_id numeric := null;
-    person_rep_id numeric := null;
+    person_id bigint := null;
+    person_rep_id bigint := null;
 begin
     errors := 0;
