+++ farms-liquibase/scripts/01_00_xx/01_00_02/06/ddl/routines/farms_import_pkg.update_status_non_autonomous.sql
@@ -1,5 +1,5 @@
 create or replace procedure farms_import_pkg.update_status_non_autonomous(
-   in in_version_id numeric,
+   in in_version_id bigint,
    in in_msg varchar
 )
 language plpgsql
