+++ farms-liquibase/scripts/01_00_xx/01_00_02/07/ddl/routines/farms_version_pkg.start_import.sql
@@ -1,5 +1,5 @@
 create or replace procedure farms_version_pkg.start_import(
-   in in_version_id numeric,
+   in in_version_id bigint,
    in in_user varchar
 )
 language plpgsql
